+++ packages/gamut-labs/src/GlobalPage/index.tsx
@@ -1,16 +1,17 @@
 import {
   AppWrapper,
+  PageWrapper,
   SkipToContent,
   SkipToContentTarget,
 } from '@codecademy/gamut';
-import { theme } from '@codecademy/gamut-styles';
+import { BackgroundProps } from '@codecademy/gamut-styles';
 import React from 'react';
 
 import { GlobalFooter, GlobalFooterProps } from '../GlobalFooter';
 import { GlobalHeader, GlobalHeaderProps } from '../GlobalHeader';
 
 export type GlobalPageProps = {
-  backgroundColor?: keyof typeof theme.colors;
+  backgroundColor?: BackgroundProps['bg'];
 
   /**
    * Element type to render around the children.
@@ -36,7 +37,7 @@ export type GlobalPageProps = {
 const defaultSkipToContentId = 'page-skip-to-content-target';
 
 export const GlobalPage: React.FC<GlobalPageProps> = ({
-  backgroundColor,
+  backgroundColor = 'background',
   children,
   contentAs = 'div',
   footer,
@@ -44,12 +45,12 @@ export const GlobalPage: React.FC<GlobalPageProps> = ({
   skipToContentId,
 }) => {
   return (
-    <AppWrapper backgroundColor={backgroundColor}>
+    <PageWrapper bg={backgroundColor}>
       <SkipToContent contentId={skipToContentId || defaultSkipToContentId} />
       <GlobalHeader {...header} />
       {!skipToContentId && <SkipToContentTarget id={defaultSkipToContentId} />}
       <AppWrapper as={contentAs}>{children}</AppWrapper>
       <GlobalFooter {...footer} />
-    </AppWrapper>
+    </PageWrapper>
   );
 };